Start protocol tests
