+++ lib/src/lib.rs
@@ -9,6 +9,9 @@ use std::io::{Read, Write};
 use sha2::{Digest, Sha256};
 use spake2::{Ed25519Group, Identity, Password, Spake2};
 
+#[cfg(test)]
+mod tests;
+
 mod chunks;
 use chunks::PortalChunks;
 
@@ -52,6 +55,9 @@ pub struct Portal {
     // Derived session key
     key: Option<Vec<u8>>,
 }
+
+pub type VerifyCallback = fn(&str, u64) -> bool;
+pub type ProgressCallback = fn(usize);
 
 impl Portal {
     /// Initialize a new portal request
@@ -148,15 +154,14 @@ impl Portal {
     /// // Begin sending the file
     /// portal.send_file(&mut client, "/etc/passwd", Some(progress));
     /// ```
-    pub fn send_file<W, D>(
+    pub fn send_file<W>(
         &mut self,
         peer: &mut W,
         path: &str,
-        callback: Option<D>,
+        callback: Option<ProgressCallback>,
     ) -> Result<usize, Box<dyn Error>>
     where
         W: Write,
-        D: Fn(usize),
     {
         // Check that the key exists to confirm the handshake is complete
         let key = self.key.as_ref().ok_or(NoPeer)?;
@@ -225,16 +230,14 @@ impl Portal {
     /// // Begin receiving the file
     /// portal.recv_file(&mut client, Some(verify_callback), Some(display_callback));
     /// ```
-    pub fn recv_file<R, V, D>(
+    pub fn recv_file<R>(
         &mut self,
         peer: &mut R,
-        verify: Option<V>,
-        display: Option<fn(usize)>,
+        verify: Option<VerifyCallback>,
+        display: Option<ProgressCallback>,
     ) -> Result<Metadata, Box<dyn Error>>
     where
         R: Read,
-        V: Fn(&str, u64) -> bool,
-        D: Fn(usize),
     {
         // Check that the key exists to confirm the handshake is complete
         let key = self.key.as_ref().ok_or(NoPeer)?;
@@ -319,430 +322,24 @@ impl Portal {
     }
 }
 
-#[cfg(test)]
-mod tests {
-    use crate::{errors::PortalError, Direction, Portal};
-    use hkdf::Hkdf;
-    use mockstream::SyncMockStream;
-    use rand::Rng;
-    use sha2::Sha256;
-    use std::io::{Read, Write};
-    use std::sync::atomic::{AtomicUsize, Ordering};
-    use std::sync::mpsc::channel;
-    use std::sync::Arc;
-    use std::thread;
-
-    pub struct MockTcpStream {
-        pub id: Direction,
-        pub waiting_for_write: Arc<AtomicUsize>,
-        pub readbuf: SyncMockStream,
-        pub write_done: Arc<AtomicUsize>,
-        pub writebuf: SyncMockStream,
-    }
-
-    impl Read for MockTcpStream {
-        fn read(&mut self, buf: &mut [u8]) -> Result<usize, std::io::Error> {
-            // Blocking read, wait until data is available
-            while self.waiting_for_write.load(Ordering::Relaxed) == 0 {
-                std::thread::sleep(std::time::Duration::from_millis(10));
-            }
-
-            // Perform the read
-            let res = self.readbuf.read(buf)?;
-
-            // Subtract the amount read from the atomic
-            self.waiting_for_write.fetch_sub(res, Ordering::Relaxed);
-            Ok(res)
-        }
-    }
-
-    impl Write for MockTcpStream {
-        fn write(&mut self, buf: &[u8]) -> Result<usize, std::io::Error> {
-            // Push data onto the peer's buffer
-            self.writebuf.push_bytes_to_read(buf);
-
-            // If they are blocked, signal that data is ready
-            if self.write_done.load(Ordering::Relaxed) == 0 {
-                self.write_done.store(buf.len(), Ordering::Relaxed);
-            }
-
-            // Return the amount written
-            Ok(buf.len())
-        }
-
-        fn flush(&mut self) -> Result<(), std::io::Error> {
-            self.writebuf.flush()
-        }
-    }
-
-    #[test]
-    fn handshake_suceeds() {
-        let fsize = 1337;
-        let fname = "filename".to_string();
-
-        // receiver
-        let dir = Direction::Receiver;
-        let pass = "test".to_string();
-        let mut receiver = Portal::init(dir, "id".to_string(), pass).unwrap();
-
-        // sender
-        let dir = Direction::Sender;
-        let pass = "test".to_string();
-        let mut sender = Portal::init(dir, "id".to_string(), pass).unwrap();
-
-        // Backing buffers
-        let senderbuf = SyncMockStream::new();
-        let receiverbuf = SyncMockStream::new();
-
-        // Backing bools
-        let senderbool = Arc::new(AtomicUsize::default());
-        let recvbool = Arc::new(AtomicUsize::default());
-        senderbool.store(0, Ordering::Relaxed);
-        recvbool.store(0, Ordering::Relaxed);
-
-        // Wrap in Mock type
-        let mut senderstream = MockTcpStream {
-            id: Direction::Sender,
-            waiting_for_write: senderbool.clone(),
-            readbuf: senderbuf.clone(),
-            write_done: recvbool.clone(),
-            writebuf: receiverbuf.clone(),
-        };
-
-        let mut recieverstream = MockTcpStream {
-            id: Direction::Receiver,
-            waiting_for_write: recvbool,
-            readbuf: receiverbuf,
-            write_done: senderbool,
-            writebuf: senderbuf,
-        };
-
-        let sender_thread = thread::spawn(move || {
-            sender.handshake(&mut senderstream).unwrap();
-        });
-
-        receiver.handshake(&mut recieverstream).unwrap();
-
-        sender_thread.join().unwrap();
-
-        // we need a key to be able to encrypt & decrypt
-
-        // Send metadata
-        //sender.write_metadata_to(&mut stream).unwrap();
-
-        // Recv metadata
-        //receiver.read_metadata_from(&mut stream).unwrap();
-
-        /* Verify both peers now share the same metadata
-        assert_eq!(fsize, receiver.get_file_size());
-        assert_eq!(fname, receiver.get_file_name().unwrap());
-        assert_eq!(
-            sender.get_file_name().unwrap(),
-            receiver.get_file_name().unwrap()
-        );
-        assert_eq!(sender.get_file_size(), receiver.get_file_size()); */
-    }
-    /*
-
-    #[test]
-    fn fail_decrypt_metadata() {
-        let fsize = 1337;
-        let fname = "filename".to_string();
-
-        // receiver
-        let dir = Direction::Receiver;
-        let pass = "test".to_string();
-        let (mut receiver, receiver_msg) = Portal::init(dir, "id".to_string(), pass, None);
-
-        // sender
-        let dir = Direction::Sender;
-        let pass = "test".to_string();
-        let (mut sender, sender_msg) =
-            Portal::init(dir, "id".to_string(), pass, Some(fname.clone()));
-        sender.set_file_size(fsize);
-
-        // we need a key to be able to encrypt & decrypt
-        receiver.derive_key(sender_msg.as_slice()).unwrap();
-        sender.derive_key(receiver_msg.as_slice()).unwrap();
-
-        // Mock channel
-        let mut stream = MockTcpStream {
-            data: Vec::with_capacity(crate::CHUNK_SIZE),
-        };
-
-        // Send garbage state
-        let mut garbage = bincode::serialize(&StateMetadata::default()).unwrap();
-        garbage.extend_from_slice(&bincode::serialize(&vec![0u8]).unwrap());
-        stream.write_all(&garbage).unwrap();
-
-        // Verify error is BadState
-        let res = receiver.read_metadata_from(&mut stream);
-        assert!(res.is_err());
-        let _ = res.map_err(|e| match e.downcast_ref::<PortalError>() {
-            Some(PortalError::BadState) => anyhow::Ok(()),
-            _ => panic!("Unexpected error"),
-        });
-
-        // Send garbage metadata
-        let state = StateMetadata {
-            nonce: rand::thread_rng().gen::<[u8; 12]>().to_vec(),
-            tag: rand::thread_rng().gen::<[u8; 16]>().to_vec(),
-        };
-        let mut garbage = bincode::serialize(&state).unwrap();
-        garbage.extend_from_slice(&bincode::serialize(&vec![0u8]).unwrap());
-        stream.write_all(&garbage).unwrap();
-
-        // Verify error is DecryptError
-        let res = receiver.read_metadata_from(&mut stream);
-        assert!(res.is_err());
-        let _ = res.map_err(|e| match e.downcast_ref::<PortalError>() {
-            Some(PortalError::DecryptError) => anyhow::Ok(()),
-            _ => panic!("Unexpected error"),
-        });
-    }
-
-    #[test]
-    fn key_derivation() {
-        // receiver
-        let dir = Direction::Receiver;
-        let pass = "test".to_string();
-        let (mut receiver, receiver_msg) = Portal::init(dir, "id".to_string(), pass, None);
-
-        // sender
-        let dir = Direction::Sender;
-        let pass = "test".to_string();
-        let (mut sender, sender_msg) = Portal::init(dir, "id".to_string(), pass, None);
-
-        receiver.derive_key(sender_msg.as_slice()).unwrap();
-        sender.derive_key(receiver_msg.as_slice()).unwrap();
-
-        assert_eq!(receiver.key, sender.key);
-    }
-
-    #[test]
-    fn key_confirmation() {
-        let mut receiver_side = MockTcpStream {
-            data: Vec::with_capacity(crate::CHUNK_SIZE),
-        };
-
-        let mut sender_side = MockTcpStream {
-            data: Vec::with_capacity(crate::CHUNK_SIZE),
-        };
-
-        // receiver
-        let dir = Direction::Receiver;
-        let pass = "test".to_string();
-        let (mut receiver, receiver_msg) = Portal::init(dir, "id".to_string(), pass, None);
-
-        // sender
-        let dir = Direction::Sender;
-        let pass = "test".to_string();
-        let (mut sender, sender_msg) = Portal::init(dir, "id".to_string(), pass, None);
-
-        receiver.derive_key(sender_msg.as_slice()).unwrap();
-        sender.derive_key(receiver_msg.as_slice()).unwrap();
-
-        // identifiers known to each party
-        let id = receiver.get_id();
-        let sender_info = format!("{}-{}", id, "senderinfo");
-        let receiver_info = format!("{}-{}", id, "receiverinfo");
-
-        // Perform the HKDF operations
-        let h = Hkdf::<Sha256>::new(None, &sender.key.as_ref().unwrap());
-        let mut sender_confirm = [0u8; 42];
-        let mut receiver_confirm = [0u8; 42];
-        h.expand(&sender_info.as_bytes(), &mut sender_confirm)
-            .unwrap();
-        h.expand(&receiver_info.as_bytes(), &mut receiver_confirm)
-            .unwrap();
-
-        // pre-send the appropriate HKDF to each stream, simulating a peer
-        receiver_side.write(&sender_confirm).unwrap();
-        sender_side.write(&receiver_confirm).unwrap();
-
-        // each side should be able to confirm the other
-        receiver.confirm_peer(&mut receiver_side).unwrap();
-        sender.confirm_peer(&mut sender_side).unwrap();
-    }
-
-    #[test]
-    fn portal_load_file() {
-        let dir = Direction::Receiver;
-        let pass = "test".to_string();
-        let (_receiver, receiver_msg) = Portal::init(dir, "id".to_string(), pass, None);
-
-        // sender
-        let dir = Direction::Sender;
-        let pass = "test".to_string();
-        let (mut sender, _sender_msg) = Portal::init(dir, "id".to_string(), pass, None);
-
-        // Confirm
-        sender.derive_key(receiver_msg.as_slice()).unwrap();
-
-        // TODO change test file
-        let _file = sender.load_file("/etc/passwd").unwrap();
-    }
-
-    #[test]
-    fn portalfile_chunks_iterator() {
-        // receiver
-        let dir = Direction::Receiver;
-        let pass = "test".to_string();
-        let (_receiver, receiver_msg) = Portal::init(dir, "id".to_string(), pass, None);
-
-        // sender
-        let dir = Direction::Sender;
-        let pass = "test".to_string();
-        let (mut sender, _sender_msg) = Portal::init(dir, "id".to_string(), pass, None);
-
-        // Confirm
-        sender.derive_key(receiver_msg.as_slice()).unwrap();
-
-        // TODO change test file
-        let file = sender.load_file("/etc/passwd").unwrap();
-
-        let chunk_size = 10;
-        for v in file.get_chunks(chunk_size) {
-            assert!(v.len() <= chunk_size);
-        }
-
-        let chunk_size = 1024;
-        for v in file.get_chunks(chunk_size) {
-            assert!(v.len() <= chunk_size);
-        }
-    }
-
-    #[test]
-    fn portal_createfile() {
-        // receiver
-        let dir = Direction::Receiver;
-        let pass = "test".to_string();
-        let (mut receiver, receiver_msg) = Portal::init(dir, "id".to_string(), pass, None);
-
-        // sender
-        let dir = Direction::Sender;
-        let pass = "test".to_string();
-        let (mut sender, sender_msg) = Portal::init(dir, "id".to_string(), pass, None);
-
-        // Confirm
-        sender.derive_key(receiver_msg.as_slice()).unwrap();
-        receiver.derive_key(sender_msg.as_slice()).unwrap();
-
-        // TODO change test file
-        let _file_dst = receiver.create_file("/tmp/passwd", 4096).unwrap();
-    }
-
-    #[test]
-    fn portal_write_chunk() {
-        // receiver
-        let dir = Direction::Receiver;
-        let pass = "test".to_string();
-        let (mut receiver, receiver_msg) = Portal::init(dir, "id".to_string(), pass, None);
-
-        // sender
-        let dir = Direction::Sender;
-        let pass = "test".to_string();
-        let (mut sender, sender_msg) = Portal::init(dir, "id".to_string(), pass, None);
-
-        // Confirm
-        sender.derive_key(receiver_msg.as_slice()).unwrap();
-        receiver.derive_key(sender_msg.as_slice()).unwrap();
-
-        // TODO change test file
-        let file_src = sender.load_file("/etc/passwd").unwrap();
-        let mut file_dst = receiver.create_file("/tmp/passwd", 4096).unwrap();
-
-        let chunk_size = 4096;
-        for v in file_src.get_chunks(chunk_size) {
-            assert!(v.len() <= chunk_size);
-
-            // test writing chunk
-            file_dst.write_given_chunk(&v).unwrap();
-        }
-    }
-
-    #[test]
-    #[should_panic]
-    fn portal_createfile_no_peer() {
-        let dir = Direction::Sender;
-        let pass = "test".to_string();
-        let (portal, _msg) = Portal::init(dir, "id".to_string(), pass, None);
-
-        // will panic due to lack of peer
-        let _file_dst = portal.create_file("/tmp/passwd", 4096).unwrap();
-    }
-
-    #[test]
-    #[should_panic]
-    fn portal_loadfile_no_peer() {
-        let dir = Direction::Sender;
-        let pass = "test".to_string();
-        let (portal, _msg) = Portal::init(dir, "id".to_string(), pass, None);
-
-        // will panic due to lack of peer
-        let _file_src = portal.load_file("/etc/passwd").unwrap();
-    }
-
-    #[test]
-    fn test_file_trim() {
-        let file = Some("/my/path/filename.txt".to_string());
-
-        let dir = Direction::Receiver;
-        let pass = "test".to_string();
-        let (receiver, _receiver_msg) = Portal::init(dir, "id".to_string(), pass, file);
-
-        let result = receiver.get_file_name().unwrap();
-
-        assert_eq!(result, "filename.txt");
-    }
-
-    #[test]
-    fn test_compressed_edwards_size() {
-        // The exchanged message is the CompressedEdwardsY + 1 byte for the SPAKE direction
-        let edwards_point = <spake2::Ed25519Group as spake2::Group>::Element::default();
-        let compressed = edwards_point.compress();
-        let msg_size: usize = std::mem::size_of_val(&compressed) + 1;
-
-        assert_eq!(33, msg_size);
-    }
-
-    #[test]
-    fn test_getters_setters() {
-        let dir = Direction::Sender;
-        let pass = "test".to_string();
-        let (mut portal, _msg) = Portal::init(dir, "id".to_string(), pass, None);
-
-        // get/set ID
-        portal.set_id("newID".to_string());
-        assert_eq!("newID", portal.get_id());
-
-        // get/set direction
-        portal.set_direction(Direction::Receiver);
-        assert_eq!(portal.get_direction(), Direction::Receiver);
-
-        // get/set direction
-        portal.set_file_size(25);
-        assert_eq!(portal.get_file_size(), 25);
-    }
-
-    #[test]
-    fn test_serialize_deserialize() {
-        let dir = Direction::Sender;
-        let pass = "test".to_string();
-        let (portal, _msg) = Portal::init(dir, "id".to_string(), pass, None);
-
-        let ser = portal.serialize().unwrap();
-        let res = Portal::parse(&ser).unwrap();
-
-        // fields that should be the same
-        assert_eq!(res.id, portal.id);
-        assert_eq!(res.direction, portal.direction);
-        assert_eq!(res.metadata.filename, portal.metadata.filename);
-        assert_eq!(res.metadata.filesize, portal.metadata.filesize);
-
-        // fields that shouldn't have been serialized
-        assert_ne!(res.state, portal.state);
-        assert_eq!(res.state, None);
-        assert_eq!(res.key, None);
-    } */
-}
+/*
+#[test]
+fn test_serialize_deserialize() {
+    let dir = Direction::Sender;
+    let pass = "test".to_string();
+    let (portal, _msg) = Portal::init(dir, "id".to_string(), pass, None);
+
+    let ser = portal.serialize().unwrap();
+    let res = Portal::parse(&ser).unwrap();
+
+    // fields that should be the same
+    assert_eq!(res.id, portal.id);
+    assert_eq!(res.direction, portal.direction);
+    assert_eq!(res.metadata.filename, portal.metadata.filename);
+    assert_eq!(res.metadata.filesize, portal.metadata.filesize);
+
+    // fields that shouldn't have been serialized
+    assert_ne!(res.state, portal.state);
+    assert_eq!(res.state, None);
+    assert_eq!(res.key, None);
+} */
